Finish restcture and run build for clean files
